cmd/awl-tray: move tray-related code to separate file
